# Reduce allocations via go-runewidth