--- a/benchstat.txt
+++ b/benchstat.txt
@@ -1,0 +1,181 @@
+goos: darwin
+goarch: arm64
+pkg: github.com/olekukonko/tablewriter/pkg/twwarp
+cpu: Apple M2
+                       │   old.txt   │              new.txt              │
+                       │   sec/op    │   sec/op     vs base              │
+WrapString-8             112.0µ ± 1%   112.9µ ± 1%  +0.80% (p=0.026 n=6)
+WrapStringWithSpaces-8   113.4µ ± 1%   114.4µ ± 0%  +0.88% (p=0.002 n=6)
+geomean                  112.7µ        113.7µ       +0.84%
+
+                       │   old.txt    │               new.txt                │
+                       │     B/op     │     B/op      vs base                │
+WrapString-8             47.35Ki ± 0%   47.35Ki ± 0%       ~ (p=1.000 n=6) ¹
+WrapStringWithSpaces-8   52.76Ki ± 0%   52.76Ki ± 0%       ~ (p=1.000 n=6)
+geomean                  49.98Ki        49.98Ki       +0.00%
+¹ all samples are equal
+
+                       │  old.txt   │              new.txt               │
+                       │ allocs/op  │ allocs/op   vs base                │
+WrapString-8             33.00 ± 0%   33.00 ± 0%       ~ (p=1.000 n=6) ¹
+WrapStringWithSpaces-8   51.00 ± 0%   51.00 ± 0%       ~ (p=1.000 n=6) ¹
+geomean                  41.02        41.02       +0.00%
+¹ all samples are equal
+
+pkg: github.com/olekukonko/tablewriter/pkg/twwidth
+                                                    │    old.txt     │               new.txt                │
+                                                    │     sec/op     │    sec/op      vs base               │
+WidthFunction/SimpleASCII_EAfalse_NoCache-8           1280.5n ±   0%   389.8n ±   0%  -69.56% (p=0.002 n=6)
+WidthFunction/SimpleASCII_EAfalse_CacheMiss-8          631.9n ± 177%   211.5n ± 135%        ~ (p=0.056 n=6)
+WidthFunction/SimpleASCII_EAfalse_CacheHit-8           14.76n ±   1%   14.76n ±   1%        ~ (p=0.387 n=6)
+WidthFunction/SimpleASCII_EAtrue_NoCache-8            1590.5n ±   1%   676.0n ±   1%  -57.50% (p=0.002 n=6)
+WidthFunction/SimpleASCII_EAtrue_CacheMiss-8           509.8n ± 316%   210.6n ± 395%        ~ (p=0.132 n=6)
+WidthFunction/SimpleASCII_EAtrue_CacheHit-8            14.73n ±   0%   14.72n ±   1%        ~ (p=0.500 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_NoCache-8          1.936µ ±   0%   1.033µ ±   0%  -46.67% (p=0.002 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_CacheMiss-8        737.2n ± 229%   264.0n ± 415%        ~ (p=1.000 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_CacheHit-8         14.74n ±   0%   14.74n ±   1%        ~ (p=0.197 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_NoCache-8           2.258µ ±   1%   1.323µ ±   3%  -41.40% (p=0.002 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_CacheMiss-8         909.2n ± 202%   635.2n ± 159%        ~ (p=1.000 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_CacheHit-8          14.72n ±   0%   14.76n ±   1%        ~ (p=0.149 n=6)
+WidthFunction/EastAsian_EAfalse_NoCache-8             1458.5n ±   1%   748.1n ±   1%  -48.71% (p=0.002 n=6)
+WidthFunction/EastAsian_EAfalse_CacheMiss-8            514.7n ± 282%   224.9n ± 386%        ~ (p=0.132 n=6)
+WidthFunction/EastAsian_EAfalse_CacheHit-8             14.78n ±   0%   14.72n ±   0%   -0.44% (p=0.017 n=6)
+WidthFunction/EastAsian_EAtrue_NoCache-8              1621.5n ±   1%   967.2n ±   0%  -40.35% (p=0.002 n=6)
+WidthFunction/EastAsian_EAtrue_CacheMiss-8             699.1n ± 212%   220.1n ± 529%        ~ (p=0.132 n=6)
+WidthFunction/EastAsian_EAtrue_CacheHit-8              14.88n ±   1%   14.74n ±   1%   -1.01% (p=0.011 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_NoCache-8      2.140µ ±   1%   1.384µ ±   1%  -35.33% (p=0.002 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_CacheMiss-8    834.8n ± 211%   635.4n ± 168%        ~ (p=0.699 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_CacheHit-8     15.71n ±   0%   15.73n ±   1%        ~ (p=0.615 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_NoCache-8       2.309µ ±   2%   1.596µ ±   0%  -30.88% (p=0.002 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_CacheMiss-8     969.0n ± 192%   464.9n ± 328%        ~ (p=0.818 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_CacheHit-8      15.74n ±   0%   15.72n ±   0%        ~ (p=0.251 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_NoCache-8       17.751µ ±   0%   4.902µ ±   1%  -72.38% (p=0.002 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_CacheMiss-8      7.080µ ± 145%   2.269µ ± 129%        ~ (p=0.699 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_CacheHit-8       23.30n ±   0%   23.70n ±   2%        ~ (p=0.063 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_NoCache-8        22.616µ ±   1%   9.299µ ±   1%  -58.88% (p=0.002 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_CacheMiss-8       8.600µ ± 107%   4.076µ ± 126%        ~ (p=0.589 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_CacheHit-8        23.30n ±   0%   23.30n ±   1%        ~ (p=0.920 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_NoCache-8      63.50µ ±   2%   57.61µ ±   1%   -9.27% (p=0.002 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_CacheMiss-8    24.09µ ± 162%   20.63µ ± 174%        ~ (p=0.699 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_CacheHit-8     44.18n ±   0%   44.19n ±   1%        ~ (p=1.000 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_NoCache-8       67.32µ ±   1%   60.60µ ±   1%   -9.98% (p=0.002 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_CacheMiss-8     23.04µ ± 189%   22.90µ ± 159%        ~ (p=0.699 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_CacheHit-8      44.39n ±   2%   44.28n ±   1%        ~ (p=0.223 n=6)
+geomean                                                562.9n          361.2n         -35.84%
+
+                                                    │     old.txt      │                 new.txt                 │
+                                                    │       B/op       │      B/op       vs base                 │
+WidthFunction/SimpleASCII_EAfalse_NoCache-8             629.0 ±   0%       112.0 ±   0%  -82.19% (p=0.002 n=6)
+WidthFunction/SimpleASCII_EAfalse_CacheMiss-8          279.00 ± 214%       55.00 ± 204%        ~ (p=0.318 n=6)
+WidthFunction/SimpleASCII_EAfalse_CacheHit-8            0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/SimpleASCII_EAtrue_NoCache-8              629.0 ±   0%       112.0 ±   1%  -82.19% (p=0.002 n=6)
+WidthFunction/SimpleASCII_EAtrue_CacheMiss-8           187.00 ± 378%       55.00 ± 504%        ~ (p=0.394 n=6)
+WidthFunction/SimpleASCII_EAtrue_CacheHit-8             0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/ASCIIWithANSI_EAfalse_NoCache-8           701.5 ±   0%       185.0 ±   0%  -73.63% (p=0.002 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_CacheMiss-8        312.50 ± 226%       95.50 ± 359%        ~ (p=0.481 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_CacheHit-8          0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/ASCIIWithANSI_EAtrue_NoCache-8            702.0 ±   0%       185.0 ±   0%  -73.65% (p=0.002 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_CacheMiss-8          347.0 ± 192%       173.5 ± 116%        ~ (p=0.481 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_CacheHit-8           0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsian_EAfalse_NoCache-8               484.5 ±   0%       145.0 ±   0%  -70.07% (p=0.002 n=6)
+WidthFunction/EastAsian_EAfalse_CacheMiss-8            188.00 ± 321%       87.00 ± 393%        ~ (p=0.545 n=6)
+WidthFunction/EastAsian_EAfalse_CacheHit-8              0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsian_EAtrue_NoCache-8                485.0 ±   0%       145.0 ±   0%  -70.10% (p=0.002 n=6)
+WidthFunction/EastAsian_EAtrue_CacheMiss-8             261.00 ± 208%       87.00 ± 392%        ~ (p=0.318 n=6)
+WidthFunction/EastAsian_EAtrue_CacheHit-8               0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsianWithANSI_EAfalse_NoCache-8       548.0 ±   0%       193.0 ±   0%  -64.78% (p=0.002 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_CacheMiss-8     338.0 ± 199%       232.0 ± 134%        ~ (p=0.697 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_CacheHit-8      0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsianWithANSI_EAtrue_NoCache-8        549.0 ±   0%       193.0 ±   0%  -64.85% (p=0.002 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_CacheMiss-8      370.5 ± 156%       152.5 ± 269%        ~ (p=0.481 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_CacheHit-8       0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongSimpleASCII_EAfalse_NoCache-8       8.289Ki ±   0%     1.150Ki ±   0%  -86.12% (p=0.002 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_CacheMiss-8     3.637Ki ± 131%     1.031Ki ±  74%        ~ (p=0.699 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_CacheHit-8        0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongSimpleASCII_EAtrue_NoCache-8        8.296Ki ±   0%     1.151Ki ±   0%  -86.12% (p=0.002 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_CacheMiss-8      3.538Ki ±  96%     1.056Ki ±  68%        ~ (p=0.777 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_CacheHit-8         0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongASCIIWithANSI_EAfalse_NoCache-8     5.839Ki ±   0%     1.356Ki ±   0%  -76.77% (p=0.002 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_CacheMiss-8   4.375Ki ±  86%     2.762Ki ±  36%        ~ (p=0.275 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_CacheHit-8      0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongASCIIWithANSI_EAtrue_NoCache-8      5.842Ki ±   0%     1.357Ki ±   0%  -76.77% (p=0.002 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_CacheMiss-8    4.193Ki ±  94%     2.791Ki ±  34%        ~ (p=0.372 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_CacheHit-8       0.000 ±   0%       0.000 ±   0%        ~ (p=1.000 n=6) ¹
+geomean                                                              ²                   -54.93%               ²
+¹ all samples are equal
+² summaries must be >0 to compute geomean
+
+                                                    │     old.txt     │                new.txt                │
+                                                    │    allocs/op    │  allocs/op    vs base                 │
+WidthFunction/SimpleASCII_EAfalse_NoCache-8            6.000 ±   0%     3.000 ±   0%  -50.00% (p=0.002 n=6)
+WidthFunction/SimpleASCII_EAfalse_CacheMiss-8          3.000 ± 133%     1.000 ± 200%        ~ (p=0.318 n=6)
+WidthFunction/SimpleASCII_EAfalse_CacheHit-8           0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/SimpleASCII_EAtrue_NoCache-8             6.000 ±   0%     3.000 ±   0%  -50.00% (p=0.002 n=6)
+WidthFunction/SimpleASCII_EAtrue_CacheMiss-8           2.000 ± 250%     1.000 ± 300%        ~ (p=0.545 n=6)
+WidthFunction/SimpleASCII_EAtrue_CacheHit-8            0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/ASCIIWithANSI_EAfalse_NoCache-8          9.000 ±   0%     6.000 ±   0%  -33.33% (p=0.002 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_CacheMiss-8        3.500 ± 186%     1.500 ± 367%        ~ (p=0.665 n=6)
+WidthFunction/ASCIIWithANSI_EAfalse_CacheHit-8         0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/ASCIIWithANSI_EAtrue_NoCache-8           9.000 ±   0%     6.000 ±   0%  -33.33% (p=0.002 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_CacheMiss-8         3.500 ± 186%     3.000 ± 133%        ~ (p=0.589 n=6)
+WidthFunction/ASCIIWithANSI_EAtrue_CacheHit-8          0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsian_EAfalse_NoCache-8              6.000 ±   0%     3.000 ±   0%  -50.00% (p=0.002 n=6)
+WidthFunction/EastAsian_EAfalse_CacheMiss-8            2.500 ± 180%     1.000 ± 300%        ~ (p=0.470 n=6)
+WidthFunction/EastAsian_EAfalse_CacheHit-8             0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsian_EAtrue_NoCache-8               6.000 ±   0%     3.000 ±   0%  -50.00% (p=0.002 n=6)
+WidthFunction/EastAsian_EAtrue_CacheMiss-8             2.500 ± 180%     1.000 ± 300%        ~ (p=0.470 n=6)
+WidthFunction/EastAsian_EAtrue_CacheHit-8              0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsianWithANSI_EAfalse_NoCache-8      8.000 ±   0%     5.000 ±   0%  -37.50% (p=0.002 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_CacheMiss-8    3.500 ± 186%     3.500 ± 100%        ~ (p=0.827 n=6)
+WidthFunction/EastAsianWithANSI_EAfalse_CacheHit-8     0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/EastAsianWithANSI_EAtrue_NoCache-8       8.000 ±   0%     5.000 ±   0%  -37.50% (p=0.002 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_CacheMiss-8     3.500 ± 186%     2.000 ± 250%        ~ (p=0.665 n=6)
+WidthFunction/EastAsianWithANSI_EAtrue_CacheHit-8      0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongSimpleASCII_EAfalse_NoCache-8        6.000 ±   0%     3.000 ±   0%  -50.00% (p=0.002 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_CacheMiss-8      4.000 ±  75%     3.000 ±  33%        ~ (p=0.643 n=6)
+WidthFunction/LongSimpleASCII_EAfalse_CacheHit-8       0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongSimpleASCII_EAtrue_NoCache-8         6.000 ±   0%     3.000 ±   0%  -50.00% (p=0.002 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_CacheMiss-8       4.000 ±  75%     3.000 ±  33%        ~ (p=0.994 n=6)
+WidthFunction/LongSimpleASCII_EAtrue_CacheHit-8        0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongASCIIWithANSI_EAfalse_NoCache-8     12.000 ±   0%     9.000 ±   0%  -25.00% (p=0.002 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_CacheMiss-8    6.000 ± 117%     4.500 ± 122%        ~ (p=0.645 n=6)
+WidthFunction/LongASCIIWithANSI_EAfalse_CacheHit-8     0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+WidthFunction/LongASCIIWithANSI_EAtrue_NoCache-8      12.000 ±   0%     9.000 ±   0%  -25.00% (p=0.002 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_CacheMiss-8     5.500 ± 136%     5.000 ± 100%        ~ (p=0.784 n=6)
+WidthFunction/LongASCIIWithANSI_EAtrue_CacheHit-8      0.000 ±   0%     0.000 ±   0%        ~ (p=1.000 n=6) ¹
+geomean                                                             ²                 -29.59%               ²
+¹ all samples are equal
+² summaries must be >0 to compute geomean
+
+pkg: github.com/olekukonko/tablewriter/tests
+                  │   old.txt    │              new.txt               │
+                  │    sec/op    │    sec/op     vs base              │
+Blueprint-8         13.44m ± 35%   13.49m ± 42%       ~ (p=0.818 n=6)
+Ocean-8             13.85m ±  1%   13.79m ±  5%       ~ (p=0.589 n=6)
+Markdown-8          13.99m ±  6%   13.60m ±  3%       ~ (p=0.240 n=6)
+Colorized-8         12.96m ±  2%   13.01m ±  2%       ~ (p=0.485 n=6)
+StreamBlueprint-8   10.49µ ±  0%   10.50µ ±  1%       ~ (p=0.457 n=6)
+StreamOcean-8       8.108µ ±  0%   8.110µ ±  0%       ~ (p=0.686 n=6)
+geomean             1.192m         1.187m        -0.39%
+
+                  │    old.txt    │                new.txt                │
+                  │     B/op      │     B/op       vs base                │
+Blueprint-8         13.72Mi ± 32%   13.77Mi ± 39%       ~ (p=0.937 n=6)
+Ocean-8             16.20Mi ±  2%   16.04Mi ±  5%       ~ (p=0.223 n=6)
+Markdown-8          15.27Mi ±  7%   15.35Mi ±  3%       ~ (p=0.310 n=6)
+Colorized-8         12.15Mi ±  3%   12.07Mi ±  2%       ~ (p=0.394 n=6)
+StreamBlueprint-8   16.12Ki ±  0%   16.12Ki ±  0%       ~ (p=1.000 n=6) ¹
+StreamOcean-8       14.65Ki ±  0%   14.65Ki ±  0%       ~ (p=1.000 n=6) ¹
+geomean             1.450Mi         1.448Mi        -0.12%
+¹ all samples are equal
+
+                  │   old.txt    │               new.txt                │
+                  │  allocs/op   │  allocs/op    vs base                │
+Blueprint-8         323.5k ± 32%   324.7k ± 39%       ~ (p=0.937 n=6)
+Ocean-8             279.8k ±  2%   277.0k ±  5%       ~ (p=0.236 n=6)
+Markdown-8          319.3k ±  7%   321.0k ±  3%       ~ (p=0.329 n=6)
+Colorized-8         378.4k ±  3%   376.1k ±  2%       ~ (p=0.394 n=6)
+StreamBlueprint-8    300.0 ±  0%    300.0 ±  0%       ~ (p=1.000 n=6) ¹
+StreamOcean-8        206.0 ±  0%    206.0 ±  0%       ~ (p=1.000 n=6) ¹
+geomean             29.62k         29.59k        -0.12%
+¹ all samples are equal