… go.sum
… new.txt
… old.txt